openmemory-mcp/requirements.txt
# Simple Universal Memory - Minimal Dependencies
mem0ai>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0

# Web UI dependencies
flask>=2.3.0

# MCP Server dependencies (Node.js)
# Run: npm install express cors